Move version package out of internal and add Makefile for assigner
